\docType{data}
\name{garner_data}
\alias{garner_data}
\title{Sample Data for Garner Filtering Analyses}
\format{
A data frame with response time and condition information for a 2x2 Garner
filtering task.
}
\usage{
data(garner_data)
}
\description{
Example data used by \code{\link{sumstats_garner}}.
}
