% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grt_wind_fit_parallel.R
\name{grt_wind_fit_parallel}
\alias{grt_wind_fit_parallel}
\title{Fit a GRT-wIND model to data several times in parallel}
\usage{
grt_wind_fit_parallel(cmats, start_params = c(), model = "full",
  rand_pert = 0.3, control = list(), n_reps, n_cores = 0)
}
\arguments{
\item{cmats}{List of confusion matrices. Each entry in the list should contain
the 4x4 confusion matrix from one individual (see Details).}

\item{start_params}{An optional vector of parameters to start the
optimization algorithm. You can provide either the group parameters or both
group and individual parameters. It will be created automatically if not
provided (see Details).}

\item{model}{A string indicating what GRT-wIND model to run. By default is the
full model ("full"), but restricted models are also available. "PS(A)" and "PS(B)"
are models that assume PS for dimension A and dimension B, respectively. "PI" is
a model that assumes perceptual independence.  "DS(A)" and "DS(B)" are models
that assume DS for dimension A ("PS(A)") and dimension B ("PS(B)"), respectively.
"1-VAR" is a model that assumes that all variances are equal to one.}

\item{rand_pert}{Maximum value of a random perturbation added to the starting
parameters. With a value of zero, the algorithm is started exactly at
\code{start_params}. As the value of \code{rand_pert} is increased, the starting
parameters become closer to be "truly random."}

\item{control}{A list of control parameters for the \code{optim} function. See
\code{\link[stats]{optim}}.}

\item{n_reps}{Number of times the optimization algorithm should be run. Must be provided.}

\item{n_cores}{Number of cores to be used. It defaults to all available cores
minus one.}
}
\value{
An object of class "\code{grt_wind_fit}."
For more details and examples, see \code{\link{grt_wind_fit}}
}
\description{
It fits a GRT-wIND model to data by running \code{\link{grt_wind_fit}} repeated times,
each time with a different value for the starting parameters. It returns the model
with a highest log-likelihood from all the runs.
}
\details{
\code{rand_pert} must be higher than zero. For more details, see 
\code{\link{grt_wind_fit}}
}
\references{
Soto, F. A., Musgrave, R., Vucovich, L., & Ashby, F. G. (2015).
General recognition theory with individual differences: A new method for
examining perceptual and decisional interactions with an application to
face perception. \emph{Psychonomic Bulletin & Review, 22}(1), 88-111.
}

