\name{grtools-package}
\alias{grtools-package}
\alias{grtools}
\docType{package}
\title{
General recognition theory tools for the analysis of perceptual independence
}
\description{
Statistical tools from General Recognition Theory for psychophysical data analyses aimed at determining independent processing of perceptual dimensions
}
\details{
\tabular{ll}{
Package: \tab grtools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-10\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Fabian Soto, Emily Zheng

Maintainer: Fabian Soto <fabian.soto@psych.ucsb.edu>
}
\references{
Ashby, F. G., & Soto, F. A. (2015). Multidimensional signal detection theory. In J. R. Busemeyer, J. T. Townsend, Z. J. Wang, & A. Eidels (Eds.), \emph{Oxford handbook of computational and mathematical psychology} (pp. 13-34). Oxford University Press: New York, NY.

Soto, F. A., Musgrave, R., Vucovich, L., & Ashby, F. G. (2015).
  General recognition theory with individual differences: A new method for
  examining perceptual and decisional interactions with an application to
  face perception. \emph{Psychonomic Bulletin & Review, 22}(1), 88-111.
  
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package, general recognition theory, grt, mdsdt, signal detection theory, 
perceptual separability, perceptual independence, separability }
\seealso{
For applications of General Recognition Theory to perceptual categorization experiments, see \code{\href{http://cran.r-project.org/web/packages/grt/index.html}{grt}
}