% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessfix.R
\name{hessfix}
\alias{hessfix}
\title{hessfix. Find hessian with fixed parameters}
\usage{
hessfix(par, fixed, fn, ..., method.args = list())
}
\description{
specify an argument 'fixed', a vector of numerical values.
created just as in function optifix
}

