% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.grt_hm_fit.R
\name{plot.grt_hm_fit}
\alias{plot.grt_hm_fit}
\title{Plot a \code{grt_hm_fit} object}
\usage{
\method{plot}{grt_hm_fit}(model, labels = c("dim A", "dim B"),
  ellipse_width = 0.8)
}
\arguments{
\item{model}{A \code{grt_hm_fit} object}

\item{labels}{Optional names for the labels of dimensions A and B}

\item{ellipse_width}{Parameter controlling the width of the drawn ellipses
#'}
}
\description{
Plot the object returned by \code{\link{grt_hm_fit}}
}

