% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstats_garner.R
\name{sumstats_garner}
\alias{sumstats_garner}
\title{Perform a summary statistics analysis of data from a Garner filtering task.}
\usage{
sumstats_garner(trial_data, test_incorrect = F)
}
\arguments{
\item{trial_data}{Trial-by-trial data from a single participant in a Garner
filtering experiment. See "Details" for instructions on the correct format
for this data frame.}

\item{test_incorrect}{If TRUE, the function runs tests on response times for
  both correct and incorrect trials. The default is FALSE.}
}
\value{
An object of class "\code{sumstats_garner}"

  The function \code{summary} is used to obtain a summary of conclusions from
  the analysis about separability. Note that any reported violations of
  separability can be due to either violations of perceptual or decisional
  separability, which cannot be dissociated through a Garner filtering
  experiment.
}
\description{
Performs an analysis of data from a 2x2 Garner filtering experiment, based on
summary statistics (see Ashby & Maddox, 1994).
}
\details{
A 2x2 Garner filtering experiment involves stimuli that vary in two
  dimensions, each with two levels, 1 and 2. The task of the participant is
  to classify stimuli according to their level in one of these dimensions
  (the relevant dimension), while ignoring variation in the other dimension
  (the irrelevant dimension).

  There are two block types in the task. During baseline blocks, the
  irrelevant dimension is fixed to a specific level. During interference
  blocks, the irrelevant dimension is not fixed, but varies across trials.
  Both accuracy and response times are gathered during the task.

  The data from a single participant in this task should be ordered in a data
  frame with rows representing individual trials and columns with the
  following format:

  \itemize{ \item{Column 1: block type, with a value of 1 for baseline blocks
  and a value of 2 for interference blocks} \item{Column 2: Level of the
  relevant dimension, with values 1 and 2} \item{Column 3: Level of the
  irrelevant dimension, with values 1 and 2} \item{Column 4: Accuracy, with a
  value of 0 for incorrect trials and 1 for correct trials} \item{Column 5:
  Response time} }

  To see an example data frame, type \code{data(garner_data)} in the R
  console. The data will be available as a \code{data.frame} named
  \code{garner_data}
}
\examples{
# Load example data frame and see the first 10 rows
data(garner_data)
garner_data[1:10,]

# Run the analysis
garner_results <- sumstats_garner(garner_data)

# See a summary of results
summary(garner_results)

# Print to screen the details of each test
garner_results
}
\references{
Ashby, F. G., & Maddox, W. T. (1994). A response time theory of
  separability and integrality in speeded classification. \emph{Journal of
  Mathematical Psychology, 38}(4), 423-466.
}

