% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grt_wind_fit_parallel.R
\name{grt_wind_fit_parallel}
\alias{grt_wind_fit_parallel}
\title{Fit a GRT-wIND model to data several times in parallel}
\usage{
grt_wind_fit_parallel(data, start_params = c(), rand_start = F,
  rand_pert = 0.3, control = list(maxit = 1000, factr = 1e+10), n_reps,
  n_cores = 0)
}
\arguments{
\item{data}{List of confusion matrices. Each entry in the list should contain
the 4x4 confusion matrix from one individual (see Details).}

\item{start_params}{An optional vector of parameters to start the
optimization algorithm. You can provide either the group parameters or both
group and individual parameters. It will be created automatically if not
provided (see Details).}

\item{rand_pert}{Maximum value of a random perturbation added to the starting
parameters. With a value of zero, the algorithm is started exactly at
start_params. As the value of rand_pert is increase, the starting
parameters become closer to be "truly random."}

\item{control}{A list of control parameters for the optim function. See
  \code{\link[stats]{optim}}.}

\item{n_reps}{Number of times the optimization algorithm should be run. Must be provided.}

\item{n_cores}{Number of cores to be used. It defaults to all available cores
minus one.}
}
\value{
An object of class "\code{grt_wind_fit}."
For more details and examples, see \code{\link{grt_wind_fit}}
}
\description{
It fits a GRT-wIND model to data by running \code{\link{grt_wind_fit}} repeated times,
each time with a different value for the starting parameters. It returns the model
with a highest log-likelihood from all the runs.
}
\details{
\code{rand_pert} must be higher than zero. For more details, see
\code{\link{grt_wind_fit}}
}
\references{
Soto, F. A., Musgrave, R., Vucovich, L., & Ashby, F. G. (2015).
General recognition theory with individual differences: A new method for
examining perceptual and decisional interactions with an application to
face perception. \emph{Psychonomic Bulletin & Review, 22}(1), 88-111.
}

