% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.grt_wind_fit.R
\name{wald}
\alias{wald}
\title{Run Wald tests of separability and independence}
\usage{
wald(fitted_model, data, estimate_hess=T)
}
\arguments{
\item{fitted_model}{An object returned by \code{\link{grt_wind_fit}}}

\item{data}{List of confusion matrices. Each entry in the list should contain
the 4x4 confusion matrix from one individual. For a detailed description of
how to create this list, see \code{\link{grt_wind_fit}}}

\item{estimate_hess}{Indicates whether a numerical estimate of the Hessian
  should be obtained. If an estimate has been previously obtained and is
  stored in \code{fitted_model}, setting \code{estimate_hess=F} will save
  computing time.}
}
\value{
An object of class "\code{grt_wind_fit}," including information about
Wald tests.

The function \code{summary} is used to obtain a summary of the model fit to
data and the results of Wald tests.
}
\description{
Run Wald tests of perceptual separability, decisional separability, and
perceptual independence (see appendix of Soto et al., 2015), using the
maximum likelihood parameter estimates stored in a \code{grt_wind_fit}
object.
}
\examples{
# Create list with confusion matrices # In this example, we enter data from
# an experiment with 5 participants. For each participant, inside the c(...),
# enter the data from row 1 in the matrix, then from row 2, etc.
data <- list(matrix(c(100,1,9,8,10,110,7,4,31,3,80,10,54,4,52,19),nrow=4,ncol=4,byrow=TRUE))
data[[2]] <- matrix(c(122,7,0,1,1,102,1,5,3,2,111,9),nrow=4,ncol=4,byrow=TRUE)
data[[3]] <- matrix(c(107,0,5,0,0,101,1,4,3,1,113,2,0,3,1,108),nrow=4,ncol=4,byrow=TRUE)
data[[4]] <- matrix(c(122,1,0,0,1,120,0,0,0,0,118,6,0,1,6,118),nrow=4,ncol=4,byrow=TRUE)
data[[5]] <- matrix(c(89,17,6,4,4,81,8,6,14,7,86,1,11,25,17,26),nrow=4,ncol=4,byrow=TRUE)

# fit the model to data
fitted_model <- grt_wind_fit(data)

#' # run the wald tests
fitted_model <- wald(fitted_model, data)

# see the results
summary(fitted_model)
}
\references{
Soto, F. A., Musgrave, R., Vucovich, L., & Ashby, F. G. (2015).
  General recognition theory with individual differences: A new method for
  examining perceptual and decisional interactions with an application to
  face perception. \emph{Psychonomic Bulletin & Review, 22}(1), 88-111.
}

